+++ lib/dradis/plugins/configurable.rb
@@ -3,14 +3,19 @@ module Dradis::Plugins
     extend ActiveSupport::Concern
 
     module ClassMethods
-      delegate :settings, to: :instance
+      delegate :encrypted_settings, :settings, to: :instance
 
       def settings_namespace
-        @settings_namespace || plugin_name
+        @settings_namespace ||= plugin_name
+      end
+
+      def addon_encrypted_settings(namespace = nil, &block)
+        @settings_namespace = namespace
+        yield self if block_given?
       end
 
       def addon_settings(namespace = nil, &block)
-        @settings_namespace = namespace if namespace
+        @settings_namespace = namespace
         yield self if block_given?
       end
 
@@ -19,8 +24,12 @@ module Dradis::Plugins
       end
     end
 
+    def encrypted_settings
+      @encrypted_settings ||= Dradis::Plugins::Settings.new(self.class.settings_namespace, adapter: :encrypted_configuration)
+    end
+
     def settings
-      @settings ||= Dradis::Plugins::Settings.new(self.class.settings_namespace)
+      @settings ||= Dradis::Plugins::Settings.new(self.class.settings_namespace, adapter: :db)
     end
   end
 end