+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-v4.9.0 (XXX 2023)
+vX.X.X (Mmmm 2023)
   - Fix deduplication of findings
+  - Store engine settings encrypted
 
 v4.8.0 (April 2023)
   - Add support for issue and content block states